Pair error handler restoration properly
Previously, the error handlers in IconvTranscoder and MbTranscoder
were not restored on error because exception was raised.
And what is worse, MbTranscoder popped the error handler stack
even when it did not push into it.

Let’s fix the former by restoring in a finally block (not catching
the raised exception), and the latter by restoring only conditionally.