mbstring: Make encoding detection stricter
PHP 8.3 changed how source encoding detection works:
https://www.php.net/manual/en/migration83.other-changes.php#migration83.other-changes.functions.mbstring
Most locales only consider `ASCII` and `UTF-8` (see `mb_detect_order()`),
and when a byte sequence invalid in both tested encodings (such as 0x91 for ‘ in Windows-1252) is encountered,
one of them might now be chosen as the most fitting encoding.
(This is done using the heuristics introduced in PHP 8.1:
https://github.com/php/php-src/commit/28b346bc0678effe7f68339ad942be16d1e1a311)

Compare the output of the following script across PHP versions:

    <?php
    $result = hex2bin("91");
    var_dump(mb_detect_encoding($result));
    var_dump(mb_detect_encoding($result, 'auto', true));
    var_dump(mb_convert_encoding($result, 'UTF-8', 'auto'));

Let’s run the `mb_detect_encoding()` ourselves with `$strict` argument set to `true`, to ensure consistent behaviour across all PHP versions.
This might potentially cause a regression is some cases. Not sure.

Additionally, since we are now ensuring all encodings are valid, we can drop the warning capture mechanism.
It does not work on PHP ≥ 8.0 anyway, since that raises a `ValueError` instead of a warning when an invalid encoding is provided.
https://www.php.net/manual/en/function.mb-convert-encoding.php#refsect1-function.mb-convert-encoding-errors

Also adjust the confusing string in tests.

https://www.php.net/manual/en/function.mb-convert-encoding.php
https://www.php.net/manual/en/function.mb-detect-encoding.php
https://www.php.net/manual/en/function.mb-detect-order.php